requirements.txt
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
